requirements.txt
requests
django~=3.2.4
pytest~=6.2.4
pytest-pythonpath==0.7.3
pytest-django
djangorestframework~=3.12.4
djangorestframework-simplejwt
django-filter~=2.4.0